requirements.txt
python-telegram-bot 
google-generativeai
# Jika kamu berencana menggunakan file .env nantinya, tambahkan:
# python-dotenv